+++ packages/react-server/examples/next/vite.config.mts
@@ -1,33 +1,6 @@
 import next from "next/vite";
-import { defineConfig, transformWithEsbuild } from "vite";
+import { defineConfig } from "vite";
 
 export default defineConfig({
-  plugins: [
-    {
-      name: "jsx-in-js",
-      async transform(code, id, _options) {
-        if (!id.includes("/node_modules/") && id.endsWith(".js")) {
-          return transformWithEsbuild(code, id, {
-            loader: "jsx",
-            jsx: "automatic",
-          });
-        }
-      },
-    },
-    next({
-      plugins: [
-        {
-          name: "jsx-in-js",
-          async transform(code, id, _options) {
-            if (!id.includes("/node_modules/") && id.endsWith(".js")) {
-              return transformWithEsbuild(code, id, {
-                loader: "jsx",
-                jsx: "automatic",
-              });
-            }
-          },
-        },
-      ],
-    }),
-  ],
+  plugins: [next()],
 });
